move rebatch serialization tests to kernel tests
